fixed naming, styles
