--- src/components/common/Extensions/SubNavigation/Partials/DropdownContent/consts.ts
@@ -1,124 +1,0 @@
-import { defineMessages } from 'react-intl';
-
-export const MSG = defineMessages({
-  paymentContentTitle: {
-    id: `common.Extensions.SubNavigation.Partials.DropdownContent.paymentContentTitle`,
-    defaultMessage: 'CREATE NEW PAYMENTS',
-  },
-  singlePayments: {
-    id: `common.Extensions.SubNavigation.Partials.DropdownContent.singlePayments`,
-    defaultMessage: 'Single payments',
-  },
-  singlePaymentsDescription: {
-    id: `common.Extensions.SubNavigation.Partials.DropdownContent.singlePaymentsDescription`,
-    defaultMessage: 'Create a new simple payment.',
-  },
-  advancedPayments: {
-    id: `common.Extensions.SubNavigation.Partials.DropdownContent.advancedPayments`,
-    defaultMessage: 'Advanced payments',
-  },
-  advancedPaymentsDescription: {
-    id: `common.Extensions.SubNavigation.Partials.DropdownContent.advancedPaymentsDescription`,
-    defaultMessage: 'A payment with more features and options.',
-  },
-  streamingPayments: {
-    id: `common.Extensions.SubNavigation.Partials.DropdownContent.streamingPayments`,
-    defaultMessage: 'Streaming payments',
-  },
-  streamingPaymentsDescription: {
-    id: `common.Extensions.SubNavigation.Partials.DropdownContent.streamingPaymentsDescription`,
-    defaultMessage: 'Create ongoing payments.',
-  },
-  moveFunds: {
-    id: `common.Extensions.SubNavigation.Partials.DropdownContent.moveFunds`,
-    defaultMessage: 'Move funds',
-  },
-  moveFundsDescription: {
-    id: `common.Extensions.SubNavigation.Partials.DropdownContent.moveFundsDescription`,
-    defaultMessage: 'Transfer funds between teams.',
-  },
-  buttonTextPay: {
-    id: `common.Extensions.SubNavigation.Partials.DropdownContent.buttonText`,
-    defaultMessage: 'View active payments',
-  },
-  helpTextPay: {
-    id: `common.Extensions.SubNavigation.Partials.DropdownContent.helpText`,
-    defaultMessage: 'Need help with payments? <a>Learn more</a>',
-  },
-  commingSoon: {
-    id: `common.Extensions.SubNavigation.Partials.DropdownContent.commingSoon`,
-    defaultMessage: 'Coming soon',
-  },
-  manageContentTitle: {
-    id: `common.Extensions.SubNavigation.Partials.DropdownContent.decisionsContentTitle`,
-    defaultMessage: 'MANAGE COLONY',
-  },
-  manageTeams: {
-    id: `common.Extensions.SubNavigation.Partials.DropdownContent.manageTeams`,
-    defaultMessage: 'Manage Teams',
-  },
-  manageTeamsDescription: {
-    id: `common.Extensions.SubNavigation.Partials.DropdownContent.manageTeamsDescription`,
-    defaultMessage: 'View, Add, and Edit teams.',
-  },
-  manageReputation: {
-    id: `common.Extensions.SubNavigation.Partials.DropdownContent.manageReputation`,
-    defaultMessage: 'Manage Reputation',
-  },
-  manageReputationDescription: {
-    id: `common.Extensions.SubNavigation.Partials.DropdownContent.manageReputationDescription`,
-    defaultMessage: 'Award or remove reputation.',
-  },
-  managePermissions: {
-    id: `common.Extensions.SubNavigation.Partials.DropdownContent.managePermissions`,
-    defaultMessage: 'Manage Permissions',
-  },
-  managePermissionsDescription: {
-    id: `common.Extensions.SubNavigation.Partials.DropdownContent.managePermissionsDescription`,
-    defaultMessage: 'Add, change or remove permissions.',
-  },
-  organizationDetails: {
-    id: `common.Extensions.SubNavigation.Partials.DropdownContent.organizationDetails`,
-    defaultMessage: 'Organization Details',
-  },
-  organizationDetailsDescription: {
-    id: `common.Extensions.SubNavigation.Partials.DropdownContent.organizationDetailsDescription`,
-    defaultMessage: 'Add or update the details of the DAO.',
-  },
-  buttonTextManage: {
-    id: `common.Extensions.SubNavigation.Partials.DropdownContent.buttonText`,
-    defaultMessage: 'View admin area',
-  },
-  helpTextManage: {
-    id: `common.Extensions.SubNavigation.Partials.DropdownContent.helpText`,
-    defaultMessage: 'Need help with admin? <a>Learn more</a>',
-  },
-  decisionsContentTitle: {
-    id: `common.Extensions.SubNavigation.Partials.DropdownContent.decisionsContentTitle`,
-    defaultMessage: 'CREATE NEW DECISIONS',
-  },
-  createDecision: {
-    id: `common.Extensions.SubNavigation.Partials.DropdownContent.createDecision`,
-    defaultMessage: 'Create Decision',
-  },
-  createDecisionDescription: {
-    id: `common.Extensions.SubNavigation.Partials.DropdownContent.createDecisionDescription`,
-    defaultMessage: 'Create a new decision.',
-  },
-  simpleDiscussion: {
-    id: `common.Extensions.SubNavigation.Partials.DropdownContent.simpleDiscussion`,
-    defaultMessage: 'Simple Discussion',
-  },
-  simpleDiscussionDescription: {
-    id: `common.Extensions.SubNavigation.Partials.DropdownContent.simpleDiscussionDescription`,
-    defaultMessage: 'Discuss plans get advice from others.',
-  },
-  buttonTextDecide: {
-    id: `common.Extensions.SubNavigation.Partials.DropdownContent.buttonText`,
-    defaultMessage: 'View decisions',
-  },
-  helpTextDecide: {
-    id: `common.Extensions.SubNavigation.Partials.DropdownContent.helpText`,
-    defaultMessage: 'Need help with decisions? <a>Learn more</a>',
-  },
-});
