+++ src/components/common/Extensions/SubNavigation/partials/DropdownContent/consts.ts
@@ -1,0 +1,124 @@
+import { defineMessages } from 'react-intl';
+
+export const MSG = defineMessages({
+  paymentContentTitle: {
+    id: `common.Extensions.SubNavigation.partials.DropdownContent.paymentContentTitle`,
+    defaultMessage: 'CREATE NEW PAYMENTS',
+  },
+  singlePayments: {
+    id: `common.Extensions.SubNavigation.partials.DropdownContent.singlePayments`,
+    defaultMessage: 'Single payments',
+  },
+  singlePaymentsDescription: {
+    id: `common.Extensions.SubNavigation.partials.DropdownContent.singlePaymentsDescription`,
+    defaultMessage: 'Create a new simple payment.',
+  },
+  advancedPayments: {
+    id: `common.Extensions.SubNavigation.partials.DropdownContent.advancedPayments`,
+    defaultMessage: 'Advanced payments',
+  },
+  advancedPaymentsDescription: {
+    id: `common.Extensions.SubNavigation.partials.DropdownContent.advancedPaymentsDescription`,
+    defaultMessage: 'A payment with more features and options.',
+  },
+  streamingPayments: {
+    id: `common.Extensions.SubNavigation.partials.DropdownContent.streamingPayments`,
+    defaultMessage: 'Streaming payments',
+  },
+  streamingPaymentsDescription: {
+    id: `common.Extensions.SubNavigation.partials.DropdownContent.streamingPaymentsDescription`,
+    defaultMessage: 'Create ongoing payments.',
+  },
+  moveFunds: {
+    id: `common.Extensions.SubNavigation.partials.DropdownContent.moveFunds`,
+    defaultMessage: 'Move funds',
+  },
+  moveFundsDescription: {
+    id: `common.Extensions.SubNavigation.partials.DropdownContent.moveFundsDescription`,
+    defaultMessage: 'Transfer funds between teams.',
+  },
+  buttonTextPay: {
+    id: `common.Extensions.SubNavigation.partials.DropdownContent.buttonText`,
+    defaultMessage: 'View active payments',
+  },
+  helpTextPay: {
+    id: `common.Extensions.SubNavigation.partials.DropdownContent.helpText`,
+    defaultMessage: 'Need help with payments? <a>Learn more</a>',
+  },
+  commingSoon: {
+    id: `common.Extensions.SubNavigation.partials.DropdownContent.commingSoon`,
+    defaultMessage: 'Coming soon',
+  },
+  manageContentTitle: {
+    id: `common.Extensions.SubNavigation.partials.DropdownContent.decisionsContentTitle`,
+    defaultMessage: 'MANAGE COLONY',
+  },
+  manageTeams: {
+    id: `common.Extensions.SubNavigation.partials.DropdownContent.manageTeams`,
+    defaultMessage: 'Manage Teams',
+  },
+  manageTeamsDescription: {
+    id: `common.Extensions.SubNavigation.partials.DropdownContent.manageTeamsDescription`,
+    defaultMessage: 'View, Add, and Edit teams.',
+  },
+  manageReputation: {
+    id: `common.Extensions.SubNavigation.partials.DropdownContent.manageReputation`,
+    defaultMessage: 'Manage Reputation',
+  },
+  manageReputationDescription: {
+    id: `common.Extensions.SubNavigation.partials.DropdownContent.manageReputationDescription`,
+    defaultMessage: 'Award or remove reputation.',
+  },
+  managePermissions: {
+    id: `common.Extensions.SubNavigation.partials.DropdownContent.managePermissions`,
+    defaultMessage: 'Manage Permissions',
+  },
+  managePermissionsDescription: {
+    id: `common.Extensions.SubNavigation.partials.DropdownContent.managePermissionsDescription`,
+    defaultMessage: 'Add, change or remove permissions.',
+  },
+  organizationDetails: {
+    id: `common.Extensions.SubNavigation.partials.DropdownContent.organizationDetails`,
+    defaultMessage: 'Organization Details',
+  },
+  organizationDetailsDescription: {
+    id: `common.Extensions.SubNavigation.partials.DropdownContent.organizationDetailsDescription`,
+    defaultMessage: 'Add or update the details of the DAO.',
+  },
+  buttonTextManage: {
+    id: `common.Extensions.SubNavigation.partials.DropdownContent.buttonText`,
+    defaultMessage: 'View admin area',
+  },
+  helpTextManage: {
+    id: `common.Extensions.SubNavigation.partials.DropdownContent.helpText`,
+    defaultMessage: 'Need help with admin? <a>Learn more</a>',
+  },
+  decisionsContentTitle: {
+    id: `common.Extensions.SubNavigation.partials.DropdownContent.decisionsContentTitle`,
+    defaultMessage: 'CREATE NEW DECISIONS',
+  },
+  createDecision: {
+    id: `common.Extensions.SubNavigation.partials.DropdownContent.createDecision`,
+    defaultMessage: 'Create Decision',
+  },
+  createDecisionDescription: {
+    id: `common.Extensions.SubNavigation.partials.DropdownContent.createDecisionDescription`,
+    defaultMessage: 'Create a new decision.',
+  },
+  simpleDiscussion: {
+    id: `common.Extensions.SubNavigation.partials.DropdownContent.simpleDiscussion`,
+    defaultMessage: 'Simple Discussion',
+  },
+  simpleDiscussionDescription: {
+    id: `common.Extensions.SubNavigation.partials.DropdownContent.simpleDiscussionDescription`,
+    defaultMessage: 'Discuss plans get advice from others.',
+  },
+  buttonTextDecide: {
+    id: `common.Extensions.SubNavigation.partials.DropdownContent.buttonText`,
+    defaultMessage: 'View decisions',
+  },
+  helpTextDecide: {
+    id: `common.Extensions.SubNavigation.partials.DropdownContent.helpText`,
+    defaultMessage: 'Need help with decisions? <a>Learn more</a>',
+  },
+});
